Renaming UnitCompositionHandler -> ConverterCompositionHandler
